Refactor components to be screens
